Add auth token to client fetch function

+++ src/lib/fetch/index.ts
@@ -1,3 +1,5 @@
+import { getSessionCookie } from '$lib';
+
 export const defaultHeaders: HeadersInit = {
 	Accept: 'application/json',
 	'Content-Type': 'application/json',
@@ -5,18 +7,50 @@ export const defaultHeaders: HeadersInit = {
 
 type Fetch = typeof fetch;
 
-export async function getResource(fn: Fetch, url: string): Promise<Response> {
-	return fn(url, { headers: defaultHeaders });
+export function authHeader(sessionToken: string) {
+	return { Authorization: `Bearer ${sessionToken}` };
 }
 
-export async function postResource(fn: Fetch, url: string, body: BodyInit): Promise<Response> {
-	return fn(url, { method: 'POST', headers: defaultHeaders, body });
+export async function getResource(url: string): Promise<Response>;
+export async function getResource(url: string, fn: Fetch): Promise<Response>;
+
+export async function getResource(url: string, fn?: Fetch): Promise<Response> {
+	if (fn) {
+		return fn(url, { headers: defaultHeaders });
+	}
+	const headers = { ...defaultHeaders, ...authHeader(getSessionCookie() as string) };
+	return fetch(url, { headers });
 }
 
-export async function patchResource(fn: Fetch, url: string, body: BodyInit): Promise<Response> {
-	return fn(url, { method: 'PATCH', headers: defaultHeaders, body });
+export async function postResource(url: string, body: BodyInit): Promise<Response>;
+export async function postResource(url: string, body: BodyInit, fn: Fetch): Promise<Response>;
+
+export async function postResource(url: string, body: BodyInit, fn?: Fetch): Promise<Response> {
+	if (fn) {
+		return fn(url, { method: 'POST', headers: defaultHeaders, body });
+	}
+	const headers = { ...defaultHeaders, ...authHeader(getSessionCookie() as string) };
+	return fetch(url, { method: 'POST', headers, body });
 }
 
-export async function deleteResource(fn: Fetch, url: string): Promise<Response> {
-	return fn(url, { method: 'DELETE', headers: defaultHeaders });
+export async function patchResource(url: string, body: BodyInit): Promise<Response>;
+export async function patchResource(url: string, body: BodyInit, fn: Fetch): Promise<Response>;
+
+export async function patchResource(url: string, body: BodyInit, fn?: Fetch): Promise<Response> {
+	if (fn) {
+		return fn(url, { method: 'PATCH', headers: defaultHeaders, body });
+	}
+	const headers = { ...defaultHeaders, ...authHeader(getSessionCookie() as string) };
+	return fetch(url, { method: 'PATCH', headers, body });
+}
+
+export async function deleteResource(url: string): Promise<Response>;
+export async function deleteResource(url: string, fn: Fetch): Promise<Response>;
+
+export async function deleteResource(url: string, fn?: Fetch): Promise<Response> {
+	if (fn) {
+		return fn(url, { method: 'DELETE', headers: defaultHeaders });
+	}
+	const headers = { ...defaultHeaders, ...authHeader(getSessionCookie() as string) };
+	return fetch(url, { method: 'DELETE', headers });
 }
